Introduce native semantic layer authoring module for Ossie and refactor commands
